@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class AnimatedSwitchDemo extends StatefulWidget {
-  const AnimatedSwitchDemo({Key key}) : super(key: key);
+  const AnimatedSwitchDemo({Key? key}) : super(key: key);
 
   @override
   _AnimatedSwitchDemoState createState() => _AnimatedSwitchDemoState();
