@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 
 class CustomScrollViewLearn extends StatelessWidget {
-  const CustomScrollViewLearn({Key key}) : super(key: key);
+  const CustomScrollViewLearn({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
